refactor web app
