+++ app-src/Utilities/BackgroundWorker.cs
@@ -5,7 +5,7 @@ using Microsoft.Extensions.Hosting;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace BAMCIS.MultiAZApp.Utils
+namespace BAMCIS.MultiAZApp.Utilities
 {
     public class BackgroundWorker : BackgroundService
     {
