+++ app-src/Utilities/BaseEnvironment.cs
@@ -1,8 +1,4 @@
-using System;
-using System.IO;
-using Microsoft.Extensions.Logging;
-
-namespace BAMCIS.MultiAZApp.Utils
+namespace BAMCIS.MultiAZApp.Utilities
 {
     public abstract class BaseEnvironment : IEnvironment
     {
