+++ app-src/Utilities/ECSEnvironment.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.Extensions.Logging;
 
-namespace BAMCIS.MultiAZApp.Utils
+namespace BAMCIS.MultiAZApp.Utilities
 {
     public class ECSEnvironment : BaseEnvironment, IEnvironment
     {
