+++ app-src/Utilities/EKSEnvironment.cs
@@ -1,7 +1,7 @@
 using System;
 using Microsoft.Extensions.Logging;
 
-namespace BAMCIS.MultiAZApp.Utils
+namespace BAMCIS.MultiAZApp.Utilities
 {
     public class EKSEnvironment : BaseEnvironment, IEnvironment
     {
