+++ app-src/Utilities/IEnvironment.cs
@@ -1,4 +1,4 @@
-namespace BAMCIS.MultiAZApp.Utils
+namespace BAMCIS.MultiAZApp.Utilities
 {
     public interface IEnvironment
     {
