--- app-src/apache/multi-az-app.conf
@@ -1,15 +1,0 @@
-<VirtualHost *:*>
-    RequestHeader set "X-Forwarded-Proto" expr=%{REQUEST_SCHEME}
-</VirtualHost>
-
-<VirtualHost *:80>
-    #RemoteIPProxyProtocol On
-    #RemoteIPProxyProtocolExceptions 127.0.0.1
-    #ProxyPreserveHost On
-    ProxyPass / http://127.0.0.1:5000/
-    ProxyPassReverse / http://127.0.0.1:5000/
-    ServerName www.example.com
-    ServerAlias *.example.com
-    ErrorLog /var/log/httpd/multi-az-app-error.log
-    CustomLog /var/log/httpd/multi-az-app-access.log common
-</VirtualHost>
