+++ app-src/appsettings.Development.json
@@ -2,15 +2,7 @@
   "Logging": {
     "LogLevel": {
       "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
+      "Microsoft.AspNetCore": "Warning"
     }
-  },
-  "Kestrel": { 
-   "Endpoints": {
-     "Local": {
-      "Url": "http://localhost:5001"
-     }
-   }
   }
 }
