--- app-src/scripts/httpd-configtest.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-# SPDX-License-Identifier: Apache-2.0
-
- service httpd configtest
